+++ lib/core/src/debug.rs
@@ -2,3 +2,6 @@
 
 pub mod monitor;
 pub mod stats;
+
+pub use monitor::*;
+pub use stats::*;
