+++ lib/core/src/display/render.rs
@@ -2,5 +2,3 @@
 
 #[warn(missing_docs)]
 pub mod css3d;
-#[warn(missing_docs)]
-pub mod webgl;
