+++ lib/core/src/display/shape/primitive/def/sdf.rs
@@ -13,9 +13,9 @@ use crate::display::shape::primitive::def::class::ShapeRef;
 use crate::display::shape::primitive::shader::canvas::Canvas;
 use crate::display::shape::primitive::shader::canvas::CanvasShape;
 use crate::display::shape::primitive::shader::data::ShaderData;
-use crate::display::render::webgl::glsl::Glsl;
+use crate::system::gpu::shader::glsl::Glsl;
 
-use crate::display::render::webgl::glsl::traits::*;
+use crate::system::gpu::shader::glsl::traits::*;
