+++ lib/core/src/system/gpu/shader.rs
