+++ lib/core/src/system/gpu/shader/glsl.rs
